# api-gateway/api/docs/swagger.yaml
definitions:
  customer.Address:
    properties:
      country:
        type: string
      id:
        type: string
      owner_id:
        type: string
      street:
        type: string
    type: object
  customer.CustomerAll:
    properties:
      customers:
        items:
          $ref: '#/definitions/customer.CustomerListRes'
        type: array
    type: object
  customer.CustomerInfo:
    properties:
      addresses:
        items:
          $ref: '#/definitions/customer.Address'
        type: array
      bio:
        type: string
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      password:
        type: string
      refresh_token:
        type: string
      updated_at:
        type: string
    type: object
  customer.CustomerListRes:
    properties:
      bio:
        type: string
      email:
        type: string
      full_name:
        type: string
    type: object
  customer.CustomerResponse:
    properties:
      access_token:
        type: string
      addresses:
        items:
          $ref: '#/definitions/customer.Address'
        type: array
      bio:
        type: string
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      password:
        type: string
      refresh_token:
        type: string
      updated_at:
        type: string
    type: object
  customer.GetAdminRes:
    properties:
      access_token:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      updated_at:
        type: string
    type: object
  customer.GetModeratorRes:
    properties:
      access_token:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      updated_at:
        type: string
    type: object
  customer.LoginResponse:
    properties:
      access_token:
        type: string
      addresses:
        items:
          $ref: '#/definitions/customer.Address'
        type: array
      bio:
        type: string
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      password:
        type: string
      refresh_token:
        type: string
      updated_at:
        type: string
    type: object
...
  models.CustomerUpdateReq:
    properties:
      bio:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
    type: object
  models.Error:
    properties:
      code:
        type: integer
      description:
        type: string
      error: {}
    type: object
...
host: localhost:9090
info:
  contact:
    email: gofurovmurtazoxon@gmail.com
    name: Murtazoxon
    url: https://t.me/murtazokhon_gofurov
  description: This is exam server api server
  termsOfService: 2 term exam
  title: exam api
  version: "2.0"
paths:
  /v1/admin/login/{admin_name}/{password}:
    get:
      consumes:
      - application/json
      description: Login admin
      parameters:
      - description: admin_name
        in: path
        name: admin_name
        required: true
        type: string
      - description: password
        in: path
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.GetAdminRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Login admin
      tags:
      - Admins
  /v1/customer/{id}:
    delete:
      consumes:
      - application/json
      description: this function delting customer
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.CustomerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: this function delete
      tags:
      - Customer
  /v1/customer/get/{id}:
    get:
      consumes:
      - application/json
      description: this function select the customers posts
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.CustomerInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: this function getting the customers posts
      tags:
      - Customer
...
  /v1/customer/update:
    put:
      consumes:
      - application/json
      description: this function updating the customers
      parameters:
      - description: Update Customer
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CustomerUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.CustomerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: this function update
      tags:
      - Customer
  /v1/customers/{page}/{limit}/{order}/{search}:
    get:
      consumes:
      - application/json
      description: Customer list getting
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: Order format should be 'key-value' key->(column_name)
        in: query
        name: order
        type: string
      - description: Search format should be 'key-value'
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.CustomerAll'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Customer list getting
      tags:
      - Customer
  /v1/login/{email}/{password}:
    get:
      consumes:
      - application/json
      description: This function get login customer
      parameters:
      - description: email
        in: path
        name: email
        required: true
        type: string
      - description: password
        in: path
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Login Customer
      tags:
      - Customer
  /v1/moderator/login/{name}/{password}:
    get:
      consumes:
      - application/json
      description: Login moderator
      parameters:
      - description: name
        in: path
        name: name
        required: true
        type: string
      - description: password
        in: path
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.GetModeratorRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Login moderator
      tags:
      - Admins
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"